requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
findspark==1.3.0
httplib2==0.17.3
idna==2.9
numpy==1.18.3
oauth2==1.9.0.post1
oauthlib==3.1.0
pandas==1.0.3
psutil==5.7.0
py4j==0.10.7
PySocks==1.7.1
pyspark==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
tweepy==3.8.0
urllib3==1.25.9
